+++ app/Http/Controllers/ApiTgroupController.php
@@ -13,7 +13,7 @@ class ApiTgroupController extends Controller
 
     private function getGroup()
     {
-        $tgroups = Tgroup::where('user_id', Auth::user()->id)->get();
+        $tgroups = Tgroup::select('id', 'name')->where('user_id', Auth::user()->id)->get();
         return $tgroups;
     }
     /**
@@ -25,88 +25,61 @@ class ApiTgroupController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
-        $validate = $request->validate([
+        $item = $request->validate([
             'name' => ['required', 'max:20', 'regex:/^[\pL\s\d]+$/u'],
         ]);
 
         $tgroup = new Tgroup;
 
-        $tgroup->name = $validate['name'];
+        $tgroup->name = $item['name'];
         $tgroup->user_id = Auth::user()->id;
         $tgroup->save();
-
-        $tgroups = Tgroup::where('user_id', Auth::user()->id)->get();
 
         $user = User::find(Auth::user()->id);
         $user->task_group = $tgroup->id;
         $user->save();
 
-        return response()->json(['groups' => $tgroups, 'active' => $tgroup->id]);
+        return response()->json(['groups' => $this->getGroup(), 'active' => $tgroup->id]);
     }
+
 
     public function check(Request $request)
     {
-        try {
-            //code...
-            //throw $th;
-            $user = User::find(Auth::user()->id);
-            switch ($request['id']) {
-                case 'task':
-                    $user->task_group = 'task';
-                    break;
+        $item = $request->validate([
+            'id' => ['required']
+        ]);
 
-                default:
-                    if (Tgroup::where('id', $request['id'])->exists()) {
-                        $user->task_group = $request['id'];
-                    } else {
-                        switch ($request['id']) {
-                            case 'daily':
-                                $user->task_group = 'daily';
-                                break;
-                            case 'date':
-                                $user->task_group = 'date';
-                                break;
-                            default:
-                                $user->task_group = 'task';
-                                break;
-                        }
+        $user = User::find(Auth::user()->id);
+
+        switch ($item['id']) {
+            case 'task':
+                $user->task_group = 'task';
+                break;
+
+            default:
+                if (Tgroup::where('id', $item['id'])->exists()) {
+                    $user->task_group = $item['id'];
+                } else {
+                    switch ($item['id']) {
+                        case 'daily':
+                            $user->task_group = 'daily';
+                            break;
+                        case 'date':
+                            $user->task_group = 'date';
+                            break;
+                        default:
+                            $user->task_group = 'task';
+                            break;
                     }
-
-                    break;
-            }
-            $user->save();
-            return response()->json(['active' => $user->task_group]);
-        } catch (\Throwable $th) {
-            return response()->json(['error' => 'generic']);
+                }
+                break;
         }
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(string $id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        //
+        $user->save();
+        return response()->json(['active' => $user->task_group]);
     }
 
     /**
@@ -114,16 +87,18 @@ class ApiTgroupController extends Controller
      */
     public function update(Request $request)
     {
-        try {
-            $tGroup = Tgroup::find($request['id']);
-            if ($tGroup->user_id == Auth::user()->id) {
-                $tGroup->name = $request['name'];
-                $tGroup->save();
-            }
+        $item = $request->validate([
+            'id' => ['numeric', 'required']
+        ]);
+
+        $tGroup = Tgroup::find($item['id']);
+
+        if ($tGroup->exists() && $tGroup->user_id == Auth::user()->id) {
+            $tGroup->name = $request['name'];
+            $tGroup->save();
             return response()->json(['groups' => $this->getGroup()]);
-        } catch (\Throwable $th) {
-            return response()->json(['error' => $th->getMessage(), 'groups' => $this->getGroup()]);
         }
+        return response()->json(['error' => true, 'groups' => $this->getGroup()]);
     }
 
     /**
@@ -131,21 +106,28 @@ class ApiTgroupController extends Controller
      */
     public function destroy(Request $request)
     {
-        try {
-            //code...
-            $tgroup = Tgroup::find($request->id);
+        $item = $request->validate([
+            'id' => ['numeric', 'required']
+        ]);
 
-            if ($tgroup->user_id == Auth::user()->id) {
-                $tasks = Task::where('tgroup_id', $tgroup->id)->get();
-                foreach ($tasks as $task) {
-                    $task->delete();
-                }
-                $tgroup->delete();
+        $tgroup = Tgroup::find($item['id']);
+
+        if ($tgroup->exists() && $tgroup->user_id == Auth::user()->id) {
+            $user = User::find(Auth::user()->id);
+            if (Auth::user()->task_group == $tgroup->id) {
+                $user->task_group = 'task';
+                $user->save();
             }
-        } catch (\Throwable $th) {
-            return response()->json(['error' => 'error']);
-        }
 
-        return response()->json(['status' => 'complete', 'groups' => $this->getGroup()]);
+            $tasks = Task::where('tgroup_id', $tgroup->id)->get();
+            foreach ($tasks as $task) {
+                $task->delete();
+            }
+
+            $tgroup->delete();
+
+            return response()->json(['error' => false, 'groups' => $this->getGroup(), 'active' => $user->task_group]);
+        }
+        return response()->json(['error' => 'error']);
     }
 }
